requirements_simplified.txt
Flask==3.0.0
line-bot-sdk
beautifulsoup4==4.12.2
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dotenv==1.0.0
gunicorn==21.2.0